Type: M Desc: merge pour bug suppression icones
git-svn-id: https://subversion.renater.fr/lodel/branches/version_0_8-bugfixes-branch@4534 9f73b02b-de1c-0410-bbc7-d36e22b4c400

+++ lodel/scripts/logic/class.entities_import.php
@@ -178,7 +178,7 @@ class Entities_ImportLogic extends Entities_EditionLogic
 			$fd=opendir(SITEROOT.$dir);
 			if (!$fd) die ("ERROR: cannot open the directory $dir");
 			while ($file=readdir($fd)) {
-				if ($file{0}==".") continue;
+				if ($file{0}=="." || !preg_match("/^(img-\d+(-small\d+)?|\w+-small\d+).(jpg|gif|png)$/i", $file)) continue;
 				$file=SITEROOT.$dir."/".$file;
 				if (is_file($file)) @unlink($file);
 			}
